Move the task to its own package
